@@ -1,14 +1,12 @@
 import React from "react";
 import "./header.styles.css"
-import logo from "../../assets/logo.png";
 import { Link } from "react-router-dom";
 
 export const Header:React.FC = () => {
     return (
         <div className = "headerContainer">
             <div className = "headerLogo">
-            <img src={logo} alt="Logo"  width="32" height="32" />
-                MANEKO
+                PHX9
             </div>
             <div className = "headerRow">
                 <Link to= "/main"><div className = "headerElement">Ballance</div></Link>
